fastapi/requirements.txt
# FastAPI
fastapi[standard]
uvicorn[standard]
pydantic
pydantic-settings

# Neo4j
neo4j

# LangChain
langchain
langchain-community
langchain-core
langchain-ollama

# HTTP Client
httpx
aiohttp

# Auth
pyjwt
passlib[bcrypt]
bcrypt<4.1

# Docker
docker

# Utilities
python-dotenv
python-multipart

# Document Processing
pypdf
python-docx
pymupdf

# Testing
pytest
pytest-asyncio
pytest-cov

# Development
black
ruff
mypy
